+++ checker/jtreg/subpackages/NullUnmarkedPackage.java
@@ -4,6 +4,6 @@
  * applyToSubpackages=false, so it excludes its own package from an enclosing AnnotatedFor but not
  * a nested subpackage.
  *
- * @compile/fail/ref=NullUnmarkedPackage.out -XDrawDiagnostics -processor org.checkerframework.checker.nullness.NullnessChecker -Amode=jspecify nup/package-info.java nup/sub/package-info.java nup/sub/InSubpackage.java nup/sub/deep/Deep.java
+ * @compile/fail/ref=NullUnmarkedPackage.out -XDrawDiagnostics -processor org.checkerframework.checker.nullness.NullnessChecker -AonlyAnnotatedFor nup/package-info.java nup/sub/package-info.java nup/sub/InSubpackage.java nup/sub/deep/Deep.java
  */
 public class NullUnmarkedPackage {}