+++ checker/jtreg/subpackages/urepeatable/package-info.java
@@ -1,0 +1,5 @@
+@AnnotatedFor(value = "nullness", applyToSubpackages = true)
+@AnnotatedFor(value = "index", applyToSubpackages = true)
+package urepeatable;
+
+import org.checkerframework.framework.qual.AnnotatedFor;