+++ checker/jtreg/subpackages/urepeatable/sub/package-info.java
@@ -1,0 +1,5 @@
+@UnannotatedFor(value = "nullness", applyToSubpackages = true)
+@UnannotatedFor(value = "index", applyToSubpackages = false)
+package urepeatable.sub;
+
+import org.checkerframework.framework.qual.UnannotatedFor;