+++ checker/tests/nullness-jspecifyscope/nullmarkedpkg/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package nullmarkedpkg;
-
-import org.jspecify.annotations.NullMarked;